Add "commit" function.
It is useful to be able to hand off patches from the quilt stack to a
separate version control system. This new function does just that,
allowing patches that do not collide to be committed to a VCS, and
removed from the quilt stack.

Only subversion (and a "null", forgetful VCS) are implemented in this
version.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>